Update en_US.js
Hi, 
You idea is awesome, it have potential.

Hope you like my PR.

I'll leave a small suggestion. Please don't understand it as a demand. If you like i could help doing it.
 
Almost all data from this project are stored on `en_US.js`
So, if you pretend to keep it small , that's fine. But this structure don't scale.
If you plan to scale this. We should split it in folders, in a alphabetical order.

E.g.
```
- Nouns (folder)
  - Animal (file)
  - Monster (file)
  - Weapon (file)
- Place (folder)
  - Building (file)
  - Natural (file)
```
etc